~#9.7

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   const videos = await Video.find({}).sort({ createdAt: "desc" });
@@ -7,40 +8,54 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  try {
-    const video = await Video.findById(id);
-    return res.render("video/watch", { pageTitle: video.title, video });
-  } catch {
+  const video = await Video.findById(id).populate("owner");
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
+  return res.render("video/watch", { pageTitle: video.title, video });
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
-  try {
-    const video = await Video.findById(id);
-    return res.render("video/edit", {
-      pageTitle: `Edit: ${video.title}`,
-      video,
-    });
-  } catch {
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("video/edit", {
+    pageTitle: `Edit: ${video.title}`,
+    video,
+  });
 };
 
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
-  try {
-    await Video.findByIdAndUpdate(id, {
-      title,
-      description,
-      hashtags: Video.formatHashtags(hashtags),
-    });
-    return res.redirect(`/videos/${id}`);
-  } catch {
+  const {
+    params: { id },
+    body: { title, description, hashtags },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  await Video.findByIdAndUpdate(id, {
+    title,
+    description,
+    hashtags: Video.formatHashtags(hashtags),
+  });
+  return res.redirect(`/videos/${id}`);
 };
 
 export const getUpload = (req, res) => {
@@ -51,14 +66,21 @@ export const postUpload = async (req, res) => {
   const {
     body: { title, description, hashtags },
     file: { path },
+    session: {
+      user: { _id },
+    },
   } = req;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl: path,
       hashtags: Video.formatHashtags(hashtags),
+      owner: _id,
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("video/upload", {
@@ -69,13 +91,28 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
-  try {
-    await Video.findByIdAndDelete(id);
-    return res.redirect("/");
-  } catch {
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  await Video.findByIdAndDelete(id);
+  const user = await User.findById(_id);
+  if (!user) {
+    return res.status(404).render("404", { pageTitle: "User not found" });
+  }
+  await User.findByIdAndUpdate(_id, {
+    videos: user.videos.filter((video) => String(video) !== id),
+  });
+  return res.redirect("/");
 };
 
 export const search = async (req, res) => {
